Removes basic parser methods to its own test suite
